@@ -2,7 +2,7 @@
 
 namespace Tests\Feature;
 
-use App\User;
+use App\Models\User;
 use Tests\TestCase;
 
 class RegisterTest extends TestCase
@@ -16,14 +16,14 @@ class RegisterTest extends TestCase
             'password' => 'secret',
             'password_confirmation' => 'secret',
         ])
-        ->assertSuccessful()
-        ->assertJsonStructure(['id', 'name', 'email']);
+            ->assertSuccessful()
+            ->assertJsonStructure(['id', 'name', 'email']);
     }
 
     /** @test */
     public function can_not_register_with_existing_email()
     {
-        factory(User::class)->create(['email' => 'test@test.app']);
+        User::factory()->create(['email' => 'test@test.app']);
 
         $this->postJson('/api/register', [
             'name' => 'Test User',
@@ -31,7 +31,7 @@ class RegisterTest extends TestCase
             'password' => 'secret',
             'password_confirmation' => 'secret',
         ])
-        ->assertStatus(422)
-        ->assertJsonValidationErrors(['email']);
+            ->assertStatus(422)
+            ->assertJsonValidationErrors(['email']);
     }
 }
