@@ -14,14 +14,14 @@ export const actions = {
     }
   },
 
-  nuxtClientInit ({ commit }) {
+  nuxtClientInit ({ commit, getters }) {
     const token = Cookies.get('token')
-    if (token) {
+    if (token && !getters['auth/token']) {
       commit('auth/SET_TOKEN', token)
     }
 
     const locale = Cookies.get('locale')
-    if (locale) {
+    if (locale && !getters['lang/locale']) {
       commit('lang/SET_LOCALE', { locale })
     }
   }
