@@ -4,9 +4,9 @@ import Cookies from 'js-cookie'
 export const state = () => ({
   locale: process.env.appLocale,
   locales: {
-    'en': 'EN',
+    en: 'EN',
     'zh-CN': '中文',
-    'es': 'ES'
+    es: 'ES'
   }
 })
 
