@@ -2,26 +2,16 @@
 
 namespace Tests\Feature;
 
-use App\User;
+use App\Models\User;
 use Illuminate\Support\Facades\Hash;
 use Tests\TestCase;
 
 class SettingsTest extends TestCase
 {
-    /** @var \App\User */
-    protected $user;
-
-    public function setUp(): void
-    {
-        parent::setUp();
-
-        $this->user = factory(User::class)->create();
-    }
-
     /** @test */
     public function update_profile_info()
     {
-        $this->actingAs($this->user)
+        $this->actingAs($user = User::factory()->create())
             ->patchJson('/api/settings/profile', [
                 'name' => 'Test User',
                 'email' => 'test@test.app',
@@ -30,7 +20,7 @@ class SettingsTest extends TestCase
             ->assertJsonStructure(['id', 'name', 'email']);
 
         $this->assertDatabaseHas('users', [
-            'id' => $this->user->id,
+            'id' => $user->id,
             'name' => 'Test User',
             'email' => 'test@test.app',
         ]);
@@ -39,13 +29,13 @@ class SettingsTest extends TestCase
     /** @test */
     public function update_password()
     {
-        $this->actingAs($this->user)
+        $this->actingAs($user = User::factory()->create())
             ->patchJson('/api/settings/password', [
-                'password' => 'updated1',
-                'password_confirmation' => 'updated1',
+                'password' => 'updated',
+                'password_confirmation' => 'updated',
             ])
             ->assertSuccessful();
 
-        $this->assertTrue(Hash::check('updated1', $this->user->password));
+        $this->assertTrue(Hash::check('updated', $user->password));
     }
 }
